requirements.txt
tensorflow-cpu==2.3.0
numpy<1.19.0,>=1.16.0
pandas
tqdm
matplotlib